feat(SilverfinAPI): reorganize into internal directory
