+++ lib/api/internal/axiosFactory.js
@@ -1,7 +1,7 @@
-const { firmCredentials } = require("./firmCredentials");
+const { firmCredentials } = require("../firmCredentials");
 const axios = require("axios");
 const { consola } = require("consola");
-const pkg = require("../../package.json");
+const pkg = require("../../../package.json");
 
 class AxiosFactory {
   constructor() {}
