+++ lib/api/internal/silverfinAuthorizer.js
@@ -1,4 +1,4 @@
-const { firmCredentials } = require("../api/firmCredentials");
+const { firmCredentials } = require("../firmCredentials");
 const prompt = require("prompt-sync")({ sigint: true });
 const { AxiosFactory } = require("./axiosFactory");
 const open = require("open");
