+++ lib/api/silverfinApi.js
@@ -1,12 +1,12 @@
 const { consola } = require("consola");
-const { BaseApi } = require("./baseApi");
-const { AuthenticationApi } = require("./authenticationApi");
-const { ReconciliationTextsApi } = require("./reconciliationTextsApi");
-const { SharedPartsApi } = require("./sharedPartsApi");
-const { ExportFilesApi } = require("./exportFilesApi");
-const { AccountTemplatesApi } = require("./accountTemplatesApi");
-const { LiquidTestingApi } = require("./liquidTestingApi");
-const { CompanyDataApi } = require("./companyDataApi");
+const { BaseApi } = require("./internal/baseApi");
+const { AuthenticationApi } = require("./internal/authenticationApi");
+const { ReconciliationTextsApi } = require("./internal/reconciliationTextsApi");
+const { SharedPartsApi } = require("./internal/sharedPartsApi");
+const { ExportFilesApi } = require("./internal/exportFilesApi");
+const { AccountTemplatesApi } = require("./internal/accountTemplatesApi");
+const { LiquidTestingApi } = require("./internal/liquidTestingApi");
+const { CompanyDataApi } = require("./internal/companyDataApi");
 
 class SilverfinApi extends BaseApi {
   constructor() {
